+++ Libplanet/Store/BlockSet.cs
@@ -4,15 +4,21 @@ using System.Linq;
 using System.Security.Cryptography;
 using Libplanet.Action;
 using Libplanet.Blocks;
+using LruCacheNet;
 
 namespace Libplanet.Store
 {
     public class BlockSet<T> : BaseIndex<HashDigest<SHA256>, Block<T>>
         where T : IAction, new()
     {
+        // FIXME it should be configurable
+        private const int DefaultCacheSize = 4096;
+        private readonly LruCache<HashDigest<SHA256>, Block<T>> _cache;
+
         public BlockSet(IStore store)
             : base(store)
         {
+            _cache = new LruCache<HashDigest<SHA256>, Block<T>>(DefaultCacheSize);
         }
 
         public override ICollection<HashDigest<SHA256>> Keys =>
@@ -20,7 +26,7 @@ namespace Libplanet.Store
 
         public override ICollection<Block<T>> Values =>
             Store.IterateBlockHashes()
-                .Select(Store.GetBlock<T>)
+                .Select(GetBlock)
                 .ToList();
 
         public override int Count => (int)Store.CountBlocks();
@@ -31,7 +37,7 @@ namespace Libplanet.Store
         {
             get
             {
-                Block<T> block = Store.GetBlock<T>(key);
+                Block<T> block = GetBlock(key);
                 if (block is null)
                 {
                     throw new KeyNotFoundException(
@@ -58,6 +64,7 @@ namespace Libplanet.Store
 
                 value.Validate(DateTimeOffset.UtcNow);
                 Store.PutBlock(value);
+                _cache.Add(value.Hash, value);
             }
         }
 
@@ -74,7 +81,30 @@ namespace Libplanet.Store
 
         public override bool Remove(HashDigest<SHA256> key)
         {
-            return Store.DeleteBlock(key);
+            bool deleted = Store.DeleteBlock(key);
+
+            if (deleted)
+            {
+                _cache.Remove(key);
+            }
+
+            return deleted;
+        }
+
+        private Block<T> GetBlock(HashDigest<SHA256> key)
+        {
+            if (_cache.TryGetValue(key, out Block<T> cached))
+            {
+                return cached;
+            }
+
+            Block<T> fetched = Store.GetBlock<T>(key);
+            if (!(fetched is null))
+            {
+                _cache.Add(key, fetched);
+            }
+
+            return fetched;
         }
     }
 }